agent_service/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
google-adk>=0.1.0
requests==2.31.0
pydantic>=2.9.0
python-multipart